[State Sync] Renamed "unit_tests" to "request_manager" and removed "tests" from "on_chain_config_tests"
